Remove outdated OpenAI Go library file
The go-openai.go library file was removed due to it being outdated and no longer in use. To prevent confusion and maintain the integrity of the repository, it was deemed necessary to remove this file. New, updated files have replaced the functionalities provided by this deleted file. A relevant documentation has also been added for ease of future developments.